REF: share _partial_date_slice between PeriodIndex/DatetimeIndex (#31766)
